base stat changes, battle manager update
changed the stats on a monster to be in an array so it can be easier to traverse. changed the battle manager such that it processes the skillettes and returns damage done to the attacking monster

+++ Assets/Skill.cs
@@ -16,9 +16,9 @@ public struct SkilletteResponse {
 
 public struct d_P{
     public int duration;
-    public double percent;
+    public float percent;
 
-    public d_P(int _duration, double _percent){
+    public d_P(int _duration, float _percent){
          duration = _duration;
          percent = _percent;
     }
@@ -26,10 +26,10 @@ public struct d_P{
 
 public struct Skillette {
     public TARGET target;
-    public double[] damageScaling;
+    public float[] damageScaling;
     public Dictionary<string, d_P> debuffs;
 
-   public Skillette(TARGET _target, double[] _dScale, Dictionary<string, d_P> debuffCalc){
+   public Skillette(TARGET _target, float[] _dScale, Dictionary<string, d_P> debuffCalc){
         target = _target;
         damageScaling = _dScale;
         debuffs =  debuffCalc;
@@ -43,7 +43,7 @@ public class Skill{
     public Skill(Monster _myself){
         myMonster = _myself;
         skillettes = new Skillette[1];
-        skillettes[0] = new Skillette(TARGET.SINGLE,new double[6]{0,1,0,0,0,0}, null);
+        skillettes[0] = new Skillette(TARGET.SINGLE,new float[6]{0,1,0,0,0,0}, null);
     }
     public Skill(Monster _myself, Skillette[] _sk){
         myMonster = _myself;
